test(web): fold the worker wiring and responsiveness checks into one
Two heavy browser tests — each laying out a canvas big enough to show a freeze
— measurably raised the flake rate of the whole parallel project: three
consecutive runs failed a different unrelated test each time (live drag, the
routing menu, the markdown flow), while the same tests passed alone. That is
the load-dependent shape this session already spent its morning removing, and
adding two more sources of it is not a trade worth making for coverage that
fits in one interaction.

Now one test asks both questions of the same edit: was a layout worker
constructed, and did the frame loop keep running while it worked. Both
mutation-checked by disabling offloading — no worker, and 2 painted frames
against a bar of 5.

Two consecutive full browser-project runs green afterwards, and the full
monorepo suite green.
